I realized that the code was getting muddled up with various 'actions' which occur when logging a track, so I decided to create a little plugin mechanism.  This is my first hack at implementing this new setup..
If you are simply using this script, nothing has substantially changed
(other than the removal of the -w/--wiki command line option; that is now
entirely controlled by the ~/.trackupdaterc file.

I'll be adding a few new "Targets" for my own use (such as posting each
logged track to a twitter account) in the near future.