Added support for track artwork
I had to rewrite the ugly AppleScript blob (which is now actually
AppleScript instead of JXA).  There is a hardcoded path in there
specific to me.  At some level, this is bad form.  On the other hand, I
don't think anyone uses this software other than me, so if that bothers
you, submit a PR. ;)