add persistent track id to json, use to name cover
Music (nee iTunes) has an internal unique ID for each track, stored in
the field `peristent ID`.  I thought that might be useful in the JSON,
so I retrieved and added it.

Then I realized there was no reason to compute an md5 hash to name the
album artwork file, I could just use that persistent ID.  So I did.