standardize on 'length/trackLength'
Originally, because of the Nicecast roots of this package, the track
length field was named `time/TrackTime` throughout the codebase (as the
NowPlaying.txt file uses that nomenclature).

But as more of the plugins use times and dates, this became annoying.
So now everywhere the Nicecast-inspired `time/trackTime` was used, we
now use `length/trackLength`